// examples/pothos/src/__generated__/pothos-prisma-extended.ts
/**
 * Generated by @rachelallyson/prisma-pothos-inputs. Do not edit.
 * PrismaTypes derived from your extended Prisma client so prismaField resolve `query` types match and findMany({ ...query, ...args }) type-checks.
 * Use: import type PrismaTypes from './this-file'; and pass PrismaTypes to SchemaBuilder.
 * The module at extendedPrismaTypePath must export: export type ExtendedPrisma = typeof yourExtendedClient;
 */

import type { ExtendedPrisma } from '../db-extended.js';
import type { User, Post } from '../../generated/prisma/client.js';

/** Extracted from extended client so PrismaTypes match and findMany({ ...query, ...args }) type-checks. */
type __UserFindManyArgs = NonNullable<Parameters<ExtendedPrisma['user']['findMany']>[0]>;
type __UserFindUniqueArgs = NonNullable<Parameters<ExtendedPrisma['user']['findUnique']>[0]>;

/** Extracted from extended client so PrismaTypes match and findMany({ ...query, ...args }) type-checks. */
type __PostFindManyArgs = NonNullable<Parameters<ExtendedPrisma['post']['findMany']>[0]>;
type __PostFindUniqueArgs = NonNullable<Parameters<ExtendedPrisma['post']['findUnique']>[0]>;

export default interface PrismaTypes {
    User: {
      Name: "User";
      Shape: User;
      Include: __UserFindManyArgs['include'];
      Select: __UserFindManyArgs['select'];
      OrderBy: __UserFindManyArgs['orderBy'];
      WhereUnique: __UserFindUniqueArgs['where'];
      Where: __UserFindManyArgs['where'];
      Create: {};
      Update: {};
      RelationName: "posts";
      ListRelations: "posts";
      Relations: {
      posts: { Shape: Post[]; Name: "Post"; Nullable: false; };
    };
    };

    Post: {
      Name: "Post";
      Shape: Post;
      Include: __PostFindManyArgs['include'];
      Select: __PostFindManyArgs['select'];
      OrderBy: __PostFindManyArgs['orderBy'];
      WhereUnique: __PostFindUniqueArgs['where'];
      Where: __PostFindManyArgs['where'];
      Create: {};
      Update: {};
      RelationName: "author";
      ListRelations: never;
      Relations: {
      author: { Shape: User; Name: "User"; Nullable: false; };
    };
    };

}